+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Touchscreen support for scrolling, zooming and clicking
+- Touchscreen support for scrolling, zooming, selecting and clicking
 - Secondary device attributes escape (`CSI > 0 c`)
 - Support for colon separated SGR 38/48
 - New Ctrl+C binding to cancel search and leave vi mode